Move NameSpaceBrowser
