+++ spyder/plugins/ipythonconsole/widgets/__init__.py
@@ -14,6 +14,7 @@ Widgets for the IPython Console.
 from .control import ControlWidget, PageControlWidget
 from .debugging import DebuggingWidget
 from .help import HelpWidget
+from .namespacebrowser import NamepaceBrowserWidget
 from .figurebrowser import FigureBrowserWidget
 from .kernelconnect import KernelConnectionDialog
 from .restartdialog import ConsoleRestartDialog
